+++ l10n/de_DE.js
@@ -1,15 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Full text search - Elasticsearch Platform" : "Volltextsuche - Elasticsearch-Plattform",
-    "Index using ElasticSearch" : "Indizieren mittels ElasticSearch",
-    "Extension to the _Full text search_ app to communicate with ElasticSearch." : "Erweiterung für die _Full text search_ App um sich mit ElasticSearch zu verbinden.",
-    "Elastic Search" : "ElasticSearch",
-    "Address of the Servlet" : "Adresse des Servlets",
-    "Include your credential in case authentication is required." : "Geben Sie Ihre Anmeldedaten für den Fall, dass eine Anmeldung erforderlich ist, ein.",
-    "Index" : "Index",
-    "Name of your index." : "Name Ihres Indexes",
-    "[Advanced] Analyzer tokenizer" : "[Advanced] Analyzer tokenizer",
-    "Some language might need a specific tokenizer." : "Einige Sprachen benötigen möglicherweise ein bestimmtes Analyse-Modul (Tokenizer)."
+    "Full text search - Meilisearch Platform": "Volltextsuche - Meilisearch-Plattform",
+    "Index using Meilisearch": "Indizieren mittels Meilisearch",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "Erweiterung für die _Full text search_ App um sich mit Meilisearch zu verbinden.",
+    "Meilisearch": "Meilisearch",
+    "Index": "Index",
+    "Name of your index.": "Name Ihres Indexes",
+    "Address of the Meilisearch server": "Adresse des Meilisearch-Servers",
+    "API Key": "API-Schlüssel",
+    "API key for authentication with Meilisearch.": "API-Schlüssel für die Authentifizierung bei Meilisearch.",
+    "Unknown Meilisearch setting": "Unbekannte Meilisearch-Einstellung",
+    "Invalid Meilisearch setting": "Ungültige Meilisearch-Einstellung"
 },
-"nplurals=2; plural=(n != 1);");
+    "nplurals=2; plural=(n != 1);"
+);