+++ l10n/es_EC.js
@@ -1,14 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Full text search - Elasticsearch Platform" : "Búsqueda de texto completo: plataforma Elasticsearch",
-    "Index using ElasticSearch" : "Indexar usando Elasticsearch",
-    "Extension to the _Full text search_ app to communicate with ElasticSearch." : "Extensión de la aplicación \"Búsqueda de texto completo\" para comunicarse con Elasticsearch.",
-    "Elastic Search" : "Elasticsearch",
-    "Address of the Servlet" : "Dirección del servlet",
-    "Include your credential in case authentication is required." : "Incluye tus credenciales en caso de que se requiera autenticación.",
-    "Index" : "Índice",
-    "Name of your index." : "Nombre de tu índice.",
-    "[Advanced] Analyzer tokenizer" : "[Avanzado] Tokenizador del analizador"
+    "Full text search - Meilisearch Platform": "Búsqueda de texto completo: plataforma Meilisearch",
+    "Index using Meilisearch": "Indexar usando Meilisearch",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "Extensión de la aplicación \"Búsqueda de texto completo\" para comunicarse con Meilisearch.",
+    "Meilisearch": "Meilisearch",
+    "Index": "Índice",
+    "Name of your index.": "Nombre de tu índice.",
+    "Address of the Meilisearch server": "Dirección del servidor Meilisearch",
+    "API Key": "Clave de API",
+    "API key for authentication with Meilisearch.": "Clave de API para autenticarse con Meilisearch.",
+    "Unknown Meilisearch setting": "Ajuste de Meilisearch desconocido",
+    "Invalid Meilisearch setting": "Ajuste de Meilisearch no válido"
 },
-"nplurals=3; plural=n == 1 ? 0 : n != 0 && n % 1000000 == 0 ? 1 : 2;");
+    "nplurals=3; plural=n == 1 ? 0 : n != 0 && n % 1000000 == 0 ? 1 : 2;"
+);