+++ l10n/ga.js
@@ -1,15 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Full text search - Elasticsearch Platform" : "Cuardach téacs iomlán - Ardán Elasticsearch",
-    "Index using ElasticSearch" : "Innéacs ag baint úsáide as ElasticSearch",
-    "Extension to the _Full text search_ app to communicate with ElasticSearch." : "Síneadh leis an aip _Full text search_ chun cumarsáid a dhéanamh le ElasticSearch.",
-    "Elastic Search" : "Cuardach Leaisteacha",
-    "Address of the Servlet" : "Seoladh an Servlet",
-    "Include your credential in case authentication is required." : "Cuir do dhintiúir san áireamh ar eagla go mbeidh fíordheimhniú ag teastáil.",
-    "Index" : "Innéacs",
-    "Name of your index." : "Ainm do innéacs.",
-    "[Advanced] Analyzer tokenizer" : "[Casta] Anailíseoir tokenizer",
-    "Some language might need a specific tokenizer." : "D’fhéadfadh go mbeadh gá le comharthaí ar leith i gcás roinnt teangacha."
+    "Full text search - Meilisearch Platform": "Cuardach téacs iomlán - Ardán Meilisearch",
+    "Index using Meilisearch": "Innéacs ag baint úsáide as Meilisearch",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "Síneadh leis an aip _Full text search_ chun cumarsáid a dhéanamh le Meilisearch.",
+    "Meilisearch": "Meilisearch",
+    "Index": "Innéacs",
+    "Name of your index.": "Ainm do innéacs.",
+    "Address of the Meilisearch server": "Seoladh an fhreastalaí Meilisearch",
+    "API Key": "Eochair API",
+    "API key for authentication with Meilisearch.": "Eochair API le haghaidh fíordheimhnithe le Meilisearch.",
+    "Unknown Meilisearch setting": "Socrú Meilisearch anaithnid",
+    "Invalid Meilisearch setting": "Socrú Meilisearch neamhbhailí"
 },
-"nplurals=5; plural=(n==1 ? 0 : n==2 ? 1 : n<7 ? 2 : n<11 ? 3 : 4);");
+    "nplurals=5; plural=(n==1 ? 0 : n==2 ? 1 : n<7 ? 2 : n<11 ? 3 : 4);"
+);