+++ l10n/is.js
@@ -1,14 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Full text search - Elasticsearch Platform" : "Leit í öllum textanum - Elasticsearch-kerfið",
-    "Index using ElasticSearch" : "Gera atriðaskrá með ElasticSearch",
-    "Extension to the _Full text search_ app to communicate with ElasticSearch." : "Viðbót fyrir _Full text search_ forritið til að eiga samskipti við ElasticSearch.",
-    "Elastic Search" : "Teygjanleg leit",
-    "Address of the Servlet" : "Vistfang forritlingsins (servlet)",
-    "Include your credential in case authentication is required." : "Hafðu með auðkennin þín ef auðkenningar verður krafist.",
-    "Index" : "Atriðaskrá",
-    "Name of your index." : "Nafn á atriðaskránni þinni.",
-    "[Advanced] Analyzer tokenizer" : "[Meira] Greining teikna (Analyzer tokenizer)"
+    "Full text search - Meilisearch Platform": "Leit í öllum textanum - Meilisearch-kerfið",
+    "Index using Meilisearch": "Gera atriðaskrá með Meilisearch",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "Viðbót fyrir _Full text search_ forritið til að eiga samskipti við Meilisearch.",
+    "Meilisearch": "Meilisearch",
+    "Index": "Atriðaskrá",
+    "Name of your index.": "Nafn á atriðaskránni þinni.",
+    "Address of the Meilisearch server": "Vistfang Meilisearch-þjónsins",
+    "API Key": "API-lykill",
+    "API key for authentication with Meilisearch.": "API-lykill fyrir auðkenningu við Meilisearch.",
+    "Unknown Meilisearch setting": "Óþekkt Meilisearch-stilling",
+    "Invalid Meilisearch setting": "Ógild Meilisearch-stilling"
 },
-"nplurals=2; plural=(n % 10 != 1 || n % 100 == 11);");
+    "nplurals=2; plural=(n % 10 != 1 || n % 100 == 11);"
+);