+++ l10n/lo.js
@@ -1,15 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Full text search - Elasticsearch Platform" : "ຄົ້ນຫາຂໍ້ຄວາມເຕັມ - ແພລັດຟອມ Elasticsearch",
-    "Index using ElasticSearch" : "ສ້າງດັດຊະນີໂດຍໃຊ້ Elasticsearch",
-    "Extension to the _Full text search_ app to communicate with ElasticSearch." : "ສ່ວນເສີມຂອງແອັບ _ຄົ້ນຫາຂໍ້ຄວາມເຕັມ_ ເພື່ອສື່ສານກັບ Elasticsearch.",
-    "Elastic Search" : "ການຄົ້ນຫາແບບ Elastic",
-    "Address of the Servlet" : "ທີ່ຢູ່ຂອງ Servlet",
-    "Include your credential in case authentication is required." : "ໃສ່ຂໍ້ມູນຢືນຢັນຕົວຕົນຂອງທ່ານ ໃນກໍລະນີທີ່ຕ້ອງການການຢືນຢັນຕົວຕົນ.",
-    "Index" : "ດັດຊະນີ",
-    "Name of your index." : "ຊື່ຂອງດັດຊະນີຂອງທ່ານ.",
-    "[Advanced] Analyzer tokenizer" : "[ຂັ້ນສູງ] ຕົວຕັດຄຳສຳລັບການວິເຄາະ",
-    "Some language might need a specific tokenizer." : "ບາງພາສາອາດຈະຕ້ອງການຕົວຕັດຄຳສະເພາະ."
+    "Full text search - Meilisearch Platform": "ຄົ້ນຫາຂໍ້ຄວາມເຕັມ - ແພລັດຟອມ Meilisearch",
+    "Index using Meilisearch": "ສ້າງດັດຊະນີໂດຍໃຊ້ Meilisearch",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "ສ່ວນເສີມຂອງແອັບ _ຄົ້ນຫາຂໍ້ຄວາມເຕັມ_ ເພື່ອສື່ສານກັບ Meilisearch.",
+    "Meilisearch": "Meilisearch",
+    "Index": "ດັດຊະນີ",
+    "Name of your index.": "ຊື່ຂອງດັດຊະນີຂອງທ່ານ.",
+    "Address of the Meilisearch server": "ທີ່ຢູ່ຂອງເຊີບເວີ Meilisearch",
+    "API Key": "ກະແຈ API",
+    "API key for authentication with Meilisearch.": "ກະແຈ API ສຳລັບການຢືນຢັນຕົວຕົນກັບ Meilisearch.",
+    "Unknown Meilisearch setting": "ການຕັ້ງຄ່າ Meilisearch ທີ່ບໍ່ຮູ້ຈັກ",
+    "Invalid Meilisearch setting": "ການຕັ້ງຄ່າ Meilisearch ບໍ່ຖືກຕ້ອງ"
 },
-"nplurals=1; plural=0;");
+    "nplurals=1; plural=0;"
+);