+++ l10n/mk.js
@@ -1,7 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Full text search - Elasticsearch Platform" : "Пребарување на целосен текст - Elasticsearch платформа",
-    "Elastic Search" : "Elastic Search"
+    "Full text search - Meilisearch Platform": "Пребарување на целосен текст - Meilisearch платформа",
+    "Index using Meilisearch": "Индексирање со Meilisearch",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "Екстензија на апликацијата _Пребарување на целосен текст_ за комуникација со Meilisearch.",
+    "Meilisearch": "Meilisearch",
+    "Index": "Индекс",
+    "Name of your index.": "Име на вашиот индекс.",
+    "Address of the Meilisearch server": "Адреса на серверот Meilisearch",
+    "API Key": "API-клуч",
+    "API key for authentication with Meilisearch.": "API-клуч за автентикација со Meilisearch.",
+    "Unknown Meilisearch setting": "Непозната поставка на Meilisearch",
+    "Invalid Meilisearch setting": "Невалидна поставка на Meilisearch"
 },
-"nplurals=2; plural=(n % 10 == 1 && n % 100 != 11) ? 0 : 1;");
+    "nplurals=2; plural=(n % 10 == 1 && n % 100 != 11) ? 0 : 1;"
+);