+++ l10n/nb.js
@@ -1,14 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Full text search - Elasticsearch Platform" : "Fulltekstsøk - Elasticsearch Platform",
-    "Index using ElasticSearch" : "Indeks ved hjelp av ElasticSearch",
-    "Extension to the _Full text search_ app to communicate with ElasticSearch." : "Utvidelse til appen _Fulltekstsøk_ for å kommunisere med ElasticSearch.",
-    "Elastic Search" : "Elastisk søk",
-    "Address of the Servlet" : "Adresse til Servlet",
-    "Include your credential in case authentication is required." : "Inkluder legitimasjonen din i tilfelle autentisering er nødvendig.",
-    "Index" : "Indeks",
-    "Name of your index." : "Navnet på indeksen din.",
-    "[Advanced] Analyzer tokenizer" : "[Avansert] Analyzer tokenizer"
+    "Full text search - Meilisearch Platform": "Fulltekstsøk - Meilisearch Platform",
+    "Index using Meilisearch": "Indeks ved hjelp av Meilisearch",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "Utvidelse til appen _Fulltekstsøk_ for å kommunisere med Meilisearch.",
+    "Meilisearch": "Meilisearch",
+    "Index": "Indeks",
+    "Name of your index.": "Navnet på indeksen din.",
+    "Address of the Meilisearch server": "Adressen til Meilisearch-serveren",
+    "API Key": "API-nøkkel",
+    "API key for authentication with Meilisearch.": "API-nøkkel for autentisering med Meilisearch.",
+    "Unknown Meilisearch setting": "Ukjent Meilisearch-innstilling",
+    "Invalid Meilisearch setting": "Ugyldig Meilisearch-innstilling"
 },
-"nplurals=2; plural=(n != 1);");
+    "nplurals=2; plural=(n != 1);"
+);