+++ l10n/oc.js
@@ -1,6 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Index" : "Indèx"
+    "Full text search - Meilisearch Platform": "Recèrca de tèxte integral - Plataforma Meilisearch",
+    "Index using Meilisearch": "Indexar amb Meilisearch",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "Extension de l’aplicacion _Recèrca de tèxte integral_ per comunicar amb Meilisearch.",
+    "Meilisearch": "Meilisearch",
+    "Index": "Indèx",
+    "Name of your index.": "Nom de vòstre indèx.",
+    "Address of the Meilisearch server": "Adreça del servidor Meilisearch",
+    "API Key": "Clau API",
+    "API key for authentication with Meilisearch.": "Clau API per l’autentificacion amb Meilisearch.",
+    "Unknown Meilisearch setting": "Paramètre Meilisearch desconegut",
+    "Invalid Meilisearch setting": "Paramètre Meilisearch invalid"
 },
-"nplurals=2; plural=(n > 1);");
+    "nplurals=2; plural=(n > 1);"
+);