+++ l10n/pt_BR.js
@@ -1,15 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Full text search - Elasticsearch Platform" : "Pesquisa em texto completo - Plataforma Elasticsearch",
-    "Index using ElasticSearch" : "Indexar usando o ElasticSearch",
-    "Extension to the _Full text search_ app to communicate with ElasticSearch." : "Extensão para o aplicativo _Full text search_ se comunicar com o ElasticSearch.",
-    "Elastic Search" : "Pesquisa Elastic",
-    "Address of the Servlet" : "Endereço do Servlet",
-    "Include your credential in case authentication is required." : "Inclua sua credencial no caso da autenticação ser necessária.",
-    "Index" : "Índice",
-    "Name of your index." : "Nome do seu índice.",
-    "[Advanced] Analyzer tokenizer" : "[Advanced] Tokenizer analisador",
-    "Some language might need a specific tokenizer." : "Alguns idiomas podem precisar de um tokenizador específico."
+    "Full text search - Meilisearch Platform": "Pesquisa em texto completo - Plataforma Meilisearch",
+    "Index using Meilisearch": "Indexar usando o Meilisearch",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "Extensão para o aplicativo _Full text search_ se comunicar com o Meilisearch.",
+    "Meilisearch": "Meilisearch",
+    "Index": "Índice",
+    "Name of your index.": "Nome do seu índice.",
+    "Address of the Meilisearch server": "Endereço do servidor Meilisearch",
+    "API Key": "Chave de API",
+    "API key for authentication with Meilisearch.": "Chave de API para autenticação com o Meilisearch.",
+    "Unknown Meilisearch setting": "Configuração desconhecida do Meilisearch",
+    "Invalid Meilisearch setting": "Configuração inválida do Meilisearch"
 },
-"nplurals=3; plural=(n == 0 || n == 1) ? 0 : n != 0 && n % 1000000 == 0 ? 1 : 2;");
+    "nplurals=3; plural=(n == 0 || n == 1) ? 0 : n != 0 && n % 1000000 == 0 ? 1 : 2;"
+);