+++ l10n/sc.js
@@ -1,14 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Full text search - Elasticsearch Platform" : "Chirca de su testu integrale - Prataforma ElasticSearch",
-    "Index using ElasticSearch" : "Inditziza cun ElasticSearch",
-    "Extension to the _Full text search_ app to communicate with ElasticSearch." : "Estensione de s'aplicatzione de chirca de su testu integrale pro comunicare cun ElasticSearch.",
-    "Elastic Search" : "Elastic Search",
-    "Address of the Servlet" : "Indiritzu de su servlet",
-    "Include your credential in case authentication is required." : "Inclue is credentziales tuas s'in casu siat recherta s'autenticatzione.",
-    "Index" : "Ìnditze",
-    "Name of your index." : "Nùmene de s'ìnditze tuo.",
-    "[Advanced] Analyzer tokenizer" : "[Advanced] Generadore de token de s'analizadore"
+    "Full text search - Meilisearch Platform": "Chirca de su testu integrale - Prataforma Meilisearch",
+    "Index using Meilisearch": "Inditziza cun Meilisearch",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "Estensione de s'aplicatzione de chirca de su testu integrale pro comunicare cun Meilisearch.",
+    "Meilisearch": "Meilisearch",
+    "Index": "Ìnditze",
+    "Name of your index.": "Nùmene de s'ìnditze tuo.",
+    "Address of the Meilisearch server": "Indiritzu de su serbidore Meilisearch",
+    "API Key": "Crae API",
+    "API key for authentication with Meilisearch.": "Crae API pro s’autenticatzione cun Meilisearch.",
+    "Unknown Meilisearch setting": "Cunfiguratzione Meilisearch disconnota",
+    "Invalid Meilisearch setting": "Cunfiguratzione Meilisearch non bàlida"
 },
-"nplurals=2; plural=(n != 1);");
+    "nplurals=2; plural=(n != 1);"
+);