+++ l10n/sk.js
@@ -1,14 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Full text search - Elasticsearch Platform" : "Fulltextové vyhľadávanie - Platforma Elasticsearch",
-    "Index using ElasticSearch" : "Indexovanie pomocou Elasticsearch",
-    "Extension to the _Full text search_ app to communicate with ElasticSearch." : "Rozšírenie pre _Full text search_ app pre komunikáciu s Elasticsearch",
-    "Elastic Search" : "Elastic Search",
-    "Address of the Servlet" : "Adresa servletu",
-    "Include your credential in case authentication is required." : "Pokiaľ sa vyžaduje overenie, zadaj svoje prihlásenie.",
-    "Index" : "Index",
-    "Name of your index." : "Názov vášho indexu.",
-    "[Advanced] Analyzer tokenizer" : "[Pokročilé] Analyzátor tokenizátor"
+    "Full text search - Meilisearch Platform": "Fulltextové vyhľadávanie - Platforma Meilisearch",
+    "Index using Meilisearch": "Indexovanie pomocou Meilisearch",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "Rozšírenie pre _Full text search_ app pre komunikáciu s Meilisearch",
+    "Meilisearch": "Meilisearch",
+    "Index": "Index",
+    "Name of your index.": "Názov vášho indexu.",
+    "Address of the Meilisearch server": "Adresa servera Meilisearch",
+    "API Key": "Kľúč API",
+    "API key for authentication with Meilisearch.": "Kľúč API na overenie voči Meilisearch.",
+    "Unknown Meilisearch setting": "Neznáme nastavenie Meilisearch",
+    "Invalid Meilisearch setting": "Neplatné nastavenie Meilisearch"
 },
-"nplurals=4; plural=(n % 1 == 0 && n == 1 ? 0 : n % 1 == 0 && n >= 2 && n <= 4 ? 1 : n % 1 != 0 ? 2: 3);");
+    "nplurals=4; plural=(n % 1 == 0 && n == 1 ? 0 : n % 1 == 0 && n >= 2 && n <= 4 ? 1 : n % 1 != 0 ? 2: 3);"
+);