+++ l10n/ug.js
@@ -1,15 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Full text search - Elasticsearch Platform" : "تولۇق تېكىست ئىزدەش - ئېلاستىكىلىق ئىزدەش سۇپىسى",
-    "Index using ElasticSearch" : "ElasticSearch نى ئىشلىتىپ كۆرسەتكۈچ",
-    "Extension to the _Full text search_ app to communicate with ElasticSearch." : "ElasticSearch بىلەن ئالاقە قىلىش ئۈچۈن _ تولۇق تېكىست ئىزدەش_ دېتالىغا كېڭەيتىش.",
-    "Elastic Search" : "ئېلاستىكىلىق ئىزدەش",
-    "Address of the Servlet" : "مۇلازىمېتىرنىڭ ئادرېسى",
-    "Include your credential in case authentication is required." : "دەلىللەش تەلەپ قىلىنغان ئەھۋالدا كىنىشكىڭىزنى قوشۇڭ.",
-    "Index" : "كۆرسەتكۈچ",
-    "Name of your index." : "كۆرسەتكۈچنىڭ ئىسمى.",
-    "[Advanced] Analyzer tokenizer" : "[Advanced] ئانالىزچى بەلگە",
-    "Some language might need a specific tokenizer." : "بەزى تىللار بەلگىلىك بىر سۆزنى بوغۇمغا ئايرىغۇچقا ئېھتىياجلىق."
+    "Full text search - Meilisearch Platform": "تولۇق تېكىست ئىزدەش - Meilisearch سۇپىسى",
+    "Index using Meilisearch": "Meilisearch نى ئىشلىتىپ كۆرسەتكۈچ",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "Meilisearch بىلەن ئالاقە قىلىش ئۈچۈن _ تولۇق تېكىست ئىزدەش_ دېتالىغا كېڭەيتىش.",
+    "Meilisearch": "Meilisearch",
+    "Index": "كۆرسەتكۈچ",
+    "Name of your index.": "كۆرسەتكۈچنىڭ ئىسمى.",
+    "Address of the Meilisearch server": "Meilisearch مۇلازىمېتىرىنىڭ ئادرېسى",
+    "API Key": "API ئاچقۇچى",
+    "API key for authentication with Meilisearch.": "Meilisearch بىلەن سالاھىيەت دەلىللەش ئۈچۈن API ئاچقۇچى.",
+    "Unknown Meilisearch setting": "نامەلۇم Meilisearch تەڭشىكى",
+    "Invalid Meilisearch setting": "ئىناۋەتسىز Meilisearch تەڭشىكى"
 },
-"nplurals=2; plural=(n != 1);");
+    "nplurals=2; plural=(n != 1);"
+);