+++ l10n/zh_CN.js
@@ -1,15 +1,17 @@
 OC.L10N.register(
-    "fulltextsearch_elasticsearch",
+    "fulltextsearch_meilisearch",
     {
-    "Full text search - Elasticsearch Platform" : "全文搜索 － Elasticsearch平台",
-    "Index using ElasticSearch" : "用ElasticSearch建立索引",
-    "Extension to the _Full text search_ app to communicate with ElasticSearch." : "用于与ElasticSearch通信的 _全文搜索_ 应用扩展。",
-    "Elastic Search" : "Elastic Search",
-    "Address of the Servlet" : "Servelet地址",
-    "Include your credential in case authentication is required." : "在需要认证时提供您的凭证",
-    "Index" : "索引",
-    "Name of your index." : "您索引的名称。",
-    "[Advanced] Analyzer tokenizer" : "[高级] 分析器tokenizer",
-    "Some language might need a specific tokenizer." : "部分语言可能需要特定的tokenizer。"
+    "Full text search - Meilisearch Platform": "全文搜索 － Meilisearch平台",
+    "Index using Meilisearch": "用Meilisearch建立索引",
+    "Extension to the _Full text search_ app to communicate with Meilisearch.": "用于与Meilisearch通信的 _全文搜索_ 应用扩展。",
+    "Meilisearch": "Meilisearch",
+    "Index": "索引",
+    "Name of your index.": "您索引的名称。",
+    "Address of the Meilisearch server": "Meilisearch 服务器地址",
+    "API Key": "API 密钥",
+    "API key for authentication with Meilisearch.": "用于 Meilisearch 身份验证的 API 密钥。",
+    "Unknown Meilisearch setting": "未知的 Meilisearch 设置",
+    "Invalid Meilisearch setting": "无效的 Meilisearch 设置"
 },
-"nplurals=1; plural=0;");
+    "nplurals=1; plural=0;"
+);